refactor: 난이도 계산 로직을 CourseScore로 이동
